Language Edits
An Unconference is type of event or conference in which the agenda is created by the participants at the beginning of the day. This event philosophy and facilitation style is designed to maximise collaboration and peer to peer sharing. All participants are invited to propose and facilitate sessions and we encourage participants to attend and move between sessions as they see fit. In contrast to a more traditional event in which the agenda is set in advance by the organisers or a programme committee and sessions have a tendency to be highly structured, unconference sessions are participant driven, drawing on the expertise, skill sets and interests of the people in the room. The unconference structure aims to encourage dialog, creativity, idea exchange, and building a sense of solidarity and common purpose between participants.